requirements.txt
python-chess
openai
anthropic
python-dotenv
streamlit
watchdog
